Improving relations.
Added better support (maybe?) for relations between objects.  Instead of having to pick an attribute to use as the foreign key, if Object A1 is added as an attribute to Object A2, when A2 gets saved, it'll automatically look up A1's table name and hash/range key and store those.  There's a method that will look for all related objects and load them.  The constructor sometimes does this; you have to tell it how many layers to go in order to prevent the possibility of an infinte loop if two objects are cross-related.

That was probably confusing as hell.  Whatever.  If you have questions on how it works, read the code or ask me.