Transparently switches between object and foreign key when saving.
Now stores foreign-key references in a single attribute.  When an object gets loaded, the foreign keys get detected and loaded.  When an object gets saved, the entry in the database shows the foreign key instead of the actual object, but the in-memory object you're working with still has the actual object as its attribute.